Extract tests for NameConfiguration
